refactor: split geotiff/zarr
