+++ rio_tiler/io/__init__.py
@@ -1,6 +1,6 @@
 """rio-tiler.io"""
 
-from .base import BaseReader, MultiBandReader, MultiBaseReader
+from .base import AsyncBaseReader, BaseReader, MultiBandReader, MultiBaseReader
 from .rasterio import ImageReader, Reader
 from .stac import STACReader
 from .xarray import XarrayReader
@@ -9,6 +9,7 @@ from .xarray import XarrayReader
 COGReader = Reader
 
 __all__ = [
+    "AsyncBaseReader",
     "BaseReader",
     "MultiBandReader",
     "MultiBaseReader",
