+++ rio_tiler/io/base.py
@@ -320,6 +320,107 @@ class BaseReader(SpatialMixin, metaclass=abc.ABCMeta):
 
 
 @attr.s
+class AsyncBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
+    """Rio-tiler.io AsyncBaseReader."""
+
+    input: Any = attr.ib()
+    tms: TileMatrixSet = attr.ib(default=WEB_MERCATOR_TMS)
+
+    async def __aenter__(self):
+        """Support using with Context Managers."""
+        return self
+
+    async def __aexit__(self, exc_type, exc_value, traceback):
+        """Support using with Context Managers."""
+        pass
+
+    @abc.abstractmethod
+    async def info(self) -> Info:
+        """Return Dataset's info.
+
+        Returns:
+            rio_tiler.models.Info: Dataset info.
+
+        """
+        ...
+
+    @abc.abstractmethod
+    async def statistics(self) -> dict[str, BandStatistics]:
+        """Return bands statistics from a dataset.
+
+        Returns:
+            dict[str, rio_tiler.models.BandStatistics]: bands statistics.
+
+        """
+        ...
+
+    @abc.abstractmethod
+    async def tile(self, tile_x: int, tile_y: int, tile_z: int) -> ImageData:
+        """Read a Map tile from the Dataset.
+
+        Args:
+            tile_x (int): Tile's horizontal index.
+            tile_y (int): Tile's vertical index.
+            tile_z (int): Tile's zoom level index.
+
+        Returns:
+            rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
+
+        """
+        ...
+
+    @abc.abstractmethod
+    async def part(self, bbox: BBox, **kwargs: Any) -> ImageData:
+        """Read a Part of a Dataset.
+
+        Args:
+            bbox (tuple): Output bounds (left, bottom, right, top) in target crs.
+
+        Returns:
+            rio_tiler.models.ImageData: ImageData instance with data, mask and input spatial info.
+
+        """
+        ...
+
+    @abc.abstractmethod
+    async def preview(self) -> ImageData:
+        """Read a preview of a Dataset.
+
+        Returns:
+            rio_tiler.models.ImageData: ImageData instance with data, mask and input spatial info.
+
+        """
+        ...
+
+    @abc.abstractmethod
+    async def point(self, lon: float, lat: float) -> PointData:
+        """Read a value from a Dataset.
+
+        Args:
+            lon (float): Longitude.
+            lat (float): Latitude.
+
+        Returns:
+            list: Pixel value per bands/assets.
+
+        """
+        ...
+
+    @abc.abstractmethod
+    async def feature(self, shape: dict) -> ImageData:
+        """Read a Dataset for a GeoJSON feature.
+
+        Args:
+            shape (dict): Valid GeoJSON feature.
+
+        Returns:
+            rio_tiler.models.ImageData: ImageData instance with data, mask and input spatial info.
+
+        """
+        ...
+
+
+@attr.s
 class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
     """MultiBaseReader Reader.
